Updated xserver fontconfig xkeyboard-config libX11 pixman pthreads plink pthreads to Nov 15 2020
xserver          commit 28ed4b95e90fb225fc32cfdce16df603ab600fcd
xkeyboard-config commit d395586249ef1d35299be46f3c56005793d9ceca
libX11           commit c9c4d6efbf92ab51695e2e740319503221d68eed
pixman           commit d93ec5713821106df6d57952dcf169308c7c4f0e
fontconfig       commit b2a54dfb8fe3859cf8c7f1d1d9b43318ef2a0ca3
mesa             commit 035e21e780205bef4d922f4ae346f284397247de
putty            commit 40e648db46bae513652079f5fe0c2cc1ecb258e9
pthreads         commit 2b0a86c48a8e655838d77babb3bdc03c44e8b8fd

+++ mesalib/src/util/set.c
@@ -223,13 +223,17 @@ _mesa_set_clear(struct set *set, void (*delete_function)(struct set_entry *entry
    if (!set)
       return;
 
-   set_foreach (set, entry) {
-      if (delete_function)
+   struct set_entry *entry;
+
+   for (entry = set->table; entry != set->table + set->size; entry++) {
+      if (entry_is_present(entry) && delete_function != NULL)
          delete_function(entry);
-      entry->key = deleted_key;
+
+      entry->key = NULL;
    }
 
-   set->entries = set->deleted_entries = 0;
+   set->entries = 0;
+   set->deleted_entries = 0;
 }
 
 /**
